sqlalchemy orm implemented & sqlite db integrated

+++ lab1/schemas.py
@@ -9,7 +9,7 @@ class EntrySchema(Schema):
     id = fields.Int(dump_only=True)
     user_id = fields.Int(required=True)
     category_id = fields.Int(required=True)
-    sum = fields.Int(required=True)
+    sum = fields.Float(required=True)
 
 class UserSchema(Schema):
     id = fields.Int(dump_only=True)
